==> Cryptocurrency-Algorithmic-Trading-Python/02-Financial-Data-Analysis-with-Python-and-Pandas/02-Price-Changes-and-Financial-Returns.ipynb <==
# Price Changes and Financial Returns

**Price Changes** are more meaningful and useful than prices.

import pandas as pd
import matplotlib.pyplot as plt
pd.options.display.float_format = '{:.4f}'.format
plt.style.use("seaborn")

close = pd.read_csv("../../Assets/Data-Files/close.csv", index_col = "Date", parse_dates = ["Date"])
close

# focusing on a single instrument, in this case the microsoft (MSFT) stock
msft = close.MSFT.dropna().to_frame().copy()
msft

# change column header
msft.rename(columns = {"MSFT":"Price"}, inplace = True)
msft

# shift elements in a column by a desired number of periods
# periods = 1 shifts prices forward by one day
msft.shift(periods = 1)

# create additional column that holds yesterdays price
msft["P_lag1"] = msft.shift(periods = 1)
msft

## Calculate Absolute Price Changes (Difference)

msft["P_diff"] = msft.Price.sub(msft.P_lag1) # alternative 1
msft

# pandas provides special method to calculate differences
msft["P_diff_2"] = msft.Price.diff(periods = 1) # alternative 2

msft

# checking if the two methods to calculate the difference return the same result
msft.P_diff.equals(msft.P_diff_2)

So far absolut price differences have been calculated which are **not meaningful**.

## Calculate Relative/Percentage Price Changes (Financial Returns)

# dividing price of current day (Price) by price of day before (P_lag1)
# and subtract 1
# to get relative price change
# that show if returns are negative or positive
msft.Price.div(msft.P_lag1) - 1 # alternative 1

# use build in Python method to calculate the percentage change
msft["Returns"] = msft.Price.pct_change(periods = 1) # alternative 2
msft

# example calculation of relative change between two days
# percentage return from Sep 2nd to Sep 3rd
percentage_change = msft.loc["2014-10-03"]["Price"] / msft.loc["2014-10-02"]["Price"] - 1
percentage_change

# multiplying percentage_change with 100 returns relative change
relative_change = percentage_change * 100
relative_change # in percent

**Take Home:** Relative Price Changes (Returns) are meaningful and comparable across instruments because the percentage change is independent from the absolute price level.

# drop columns and safe file for next lectures
msft.drop(columns = ["P_lag1", "P_diff", "P_diff_2"], inplace = True)
msft

msft.to_csv("../../Assets/Data-Files/msft.csv")
