==> Cryptocurrency-Algorithmic-Trading-Python/02-Financial-Data-Analysis-with-Python-and-Pandas/04-Investment-Multiple-and-CAGR.ipynb <==
# Investment Multiple and CAGR (Compound Annual Growth Rate)

### Two alternative reward metrics that are more intuitive and easier to interpret (compared to the arithmetic mean).

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
pd.options.display.float_format = '{:.4f}'.format
plt.style.use("seaborn")

msft = pd.read_csv("../../Assets/Data-Files/msft.csv", index_col = "Date", parse_dates = ["Date"])
msft

**Investment Multiple:** Ending Value of 1 (Dollar, Euro, etc.) invested.  
Multiple = Ending Value / Initial Value

# last price of a given period divided by the first price returns the multiple
multiple = (msft.Price[-1] / msft.Price[0])
multiple # result is the ending value of one dollar invested

**Price Increase (in %)**

# calculate the increase in percent
(multiple - 1) * 100 # the invested dollar gets subtracted from the multiple

# similar (not identical) concept: Normalized Price with a Base Value 1
msft.Price / msft.Price[0] # last value is equivalent to multiple

**Drawback of Investment Multiple:** Doesn't take into account investment period (5.38 is great over seven years but not so great over 100 years). Meaningful only in conjunction with investment period (this means investment period has to be taken into account which is the benefit of the CAGR).

<font size="4" style="line-height:1.3">**Compound Annual Growth Rate (CAGR):** The (constant annual) rate of return that would be required for an investment to grow from its beginning balance to its ending balance, assuming the profits were reinvested at the end of each year of the investment's lifespan (Wikipedia).</font>

start = msft.index[0]
start

end = msft.index[-1]
end

# timestamp objects can be subtracted
td = end - start
td # returns a timedelta object

# timedelta divided by average number of years = 365.25
td_years = td.days / 365.25
td_years

# Compound Annual Growth Rate is the multiple to the power of one devided by the annual timedelta
# and then subtracted by one
cagr = multiple**(1 / td_years) - 1 #short version
cagr

# more complex calculation with pandas data frame
# (msft.Price[-1]/msft.Price[0]) calculates the multiple
# then the timedelta divided by average number of years = 365.25 minus 1 is calculated
cagr = (msft.Price[-1]/msft.Price[0])**(1/((msft.index[-1] - msft.index[0]).days / 365.25)) - 1 
cagr

# inverse calculation, calculating multiple with cagr
(1 + cagr)**(td_years) # alternative 2 to calculate multiple

**Take Home:** CAGR can be used to compare investments with different investment horizons.
