==> Cryptocurrency-Algorithmic-Trading-Python/02-Financial-Data-Analysis-with-Python-and-Pandas/05-Compound-Returns-and-Geometric-Mean-Return.ipynb <==
## Compound Returns & Geometric Mean Return

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
pd.options.display.float_format = '{:.4f}'.format
plt.style.use("seaborn")

msft = pd.read_csv("../../Assets/Data-Files/msft.csv", index_col = "Date", parse_dates = ["Date"])
msft

# 3rd alternative to calculate multiple (compounding daily returns)
# prod() method multiplies all elements
multiple = (1 + msft.Returns).prod()
multiple

# number of elements in return column
n = msft.Returns.count()
n

# geometric mean return
geo_mean = multiple**(1/n) - 1
geo_mean

# alternative #4 to calculate multiple (geometric mean)
(1 + geo_mean)**n

<font size="4" style="line-height:1.3">**Geometric Mean has similar definition as the Compound Annual Growth Rate:** The (constant daily) rate of return that would be required for an investment to grow from its beginning balance to its ending balance, assuming the profits were reinvested at the end of each day of the investment's lifespan (Wikipedia).</font>

<font size="4" style="line-height:1.3">**The concepts of Compound Returns, Compound Annual Growth Rate and Geometric Mean are closely related.**</font>

mu = msft.Returns.mean() # arithmetic mean return
mu

<font size="4" style="line-height:1.3">**The arthmetic mean return is always greater then the geometric mean return and less usefull.**</font>


# calculating multiple is not possible with arithemetic mean
(1 + mu)**n # does not return multiple
