==> Cryptocurrency-Algorithmic-Trading-Python/02-Financial-Data-Analysis-with-Python-and-Pandas/06-Discrete-vs-Continuous-Compounding.ipynb <==
# Discrete vs. Continuous Compounding

## Simple Returns vs. Logarithmic Returns (log returns)

Very often log returns are used instead of simple returns.

- favourable characteristics of log returns
- drawbacks of simple returns

Discrete vs. Continuous Compounding gives a more detailed background on log returns.

## Discrete Compounding 

<font size="4" style="line-height:1.3">**Annual Compounding:** Interests accrune once a year at the end of the year.  </font>

<font size="4" style="line-height:1.3">**Example**: Your bank offers an interst rate of **8% p.a. (stated rate)** with **annual compounding** on your savings **(USD 100)**. Calculate the **value** of your savings account **after one year** and the corresponding **effective annual interest rate**.</font>

<font size="4" style="line-height:1.3">**Interests are calculated and added to your savings account once at the end of each year.**
</font>

# your savings, present value
PV = 100
# interest rate p.a.
r = 0.08
# time range, 1 year
n = 1

# value after one year
PV * (1 + r)

# official formula for future value
FV = PV * (1 + r)**n
FV

# effective annual rate
effective_annual_rate = (FV / PV)**(1/n) - 1
effective_annual_rate

In case of annual compounding effective annual rate is identical to stated rate.

<font size="4" style="line-height:1.3">**Quarterly Compounding:** Interests accrune once a quarter at the end of the quarter.  </font>

<font size="4" style="line-height:1.3">**Example**: Your bank offers an interst rate of **8% p.a. (stated rate)** with **quarterly compounding** on your savings **(USD 100)**. Calculate the **value** of your savings account **after one year** and the corresponding **effective annual interest rate**.</font>

<font size="4" style="line-height:1.3">**Interests are calculated and added to your savings account once at the end of each quarter.**
</font>

# your savings, present value
PV = 100
# interest rate p.a.
r = 0.08
# time range, 1 year
n = 1
# number of compounding events
m = 4

# value after one year
# after on quarter we earn 2%
# interest on interests = coumpound interests
100 * 1.02 * 1.02 * 1.02 * 1.02

# official formula for future value
FV = PV * (1 + r/m)**(n*m)
FV

# effective annual rate
effective_annual_rate = (FV / PV)**(1/n) - 1
effective_annual_rate

With quarterly compounding we end up at a higher value and we also have a higher effective interest rate.

<font size="4" style="line-height:1.3">**Quarterly Compounding is favourable (everything else equal) as we earn compound interest (interest on interest).**</font>


<font size="4" style="line-height:1.3">**Monthly Compounding:** Interests accrune once a month at the end of the month.  </font>

<font size="4" style="line-height:1.3">**Example**: Your savings bank offers an interst rate of **8% p.a. (stated rate)** with **quarterly compounding** on your savings **(USD 100)**. Calculate the **value** of your savings account **after one year** and the corresponding **effective annual interest rate**.</font>

<font size="4" style="line-height:1.3">**Interests are calculated and added to your savings account once at the end of each quarter.**
</font>
